Fix cfg labelling on quantified expr with finalizers
gnatcov decision map analysis wrongly assumes that finalizers (in a
context with controlled objects) necessarily comes after the decision has
been evaluated, which is false in the case of for all / for some
expresssions; a temporary variable holding the final result is created
(in the binary code) in that case, and the check is actually done
__after__ the call to finalizers. This makes gnatcov incorrectly label
branches that do not implement the control flow.

For that reason, we will deactivate that heuristic when the condition is
a quantified expression.

TN: T310-022
Change-Id: I33f78b092d4b84516abd9ac70d2498c93fe9bcdf