gnatcov instrument: work around hidden Standard package
When instrumenting a decision, we need to convert to the Standard.Boolean
type (that the Witness function accepts), and back the original type
(that the user code accepts).

When converting to a type, gnatcov fully prefixes it. But if Standard is
hidden, as noticed on some user code, it will break as Standard.Boolean
will not anymore be a valid Ada expression (if the redefinition is, e.g.
the member of an enumeration).

To work around this issue, we host a renaming of the Standard package
under the GNATcov_RTS package, that we will use instead of Standard.

We will do the same for the System package (as we use System.Address for
MC/DC states).

TN: U701-016
Change-Id: I29cda2e1571f49578439f76fdc722641e7b4d413