Add preliminary support for Short_Circuit_And_Or pragma in instrumentation
While waiting for implementation of pragma configuration file support
in Libadalang (U624-007), add an option to specify that "and" and "or"
standard boolean operators have short-circuit semantics, and instrument
them for MCDC accordingly.

A bug in Libadalang overloading analysis (U705-016)  makes this feature
miss some "and" and "or" operators that should be considered as standard
boolean operators, and are thus not instrumented for MC/DC coverage.
This should resolve itself once the bug is corrected.

Change-Id: I88d567e708a3f6266c2ee3f2611d6c0d518739fc
TN: U621-005