gnatcov instrument: change exported symbol for the Dump_Buffer procedure
The Dump_Buffer procedure is exported as a C symbol (for future C
instrumentation capabilities). This symbol was too generic, and did not
deal well with multiple mains, one being in the closure of the other (as
this ended up generating multiple Dump_Buffers procedure with the same
symbol name).

This symbol is now prefixed both by "gnatcov_rts" and by the main name.

Change-Id: Iea1351cdcd975f80122721e5267a38429abbe367
TN: U706-006