gnatcov instrument: remove Traverse_Aspects
In practice, Traverse_One is never called on Aspect_Spec nodes, and thus
the Traverse_Aspects procedure, which is only called in Traverse_One, is
never used.

This is because in Traverse_Declarations_Or_Statements, we call
Traverse_One only on statements, declarations or groups of these (e.g.
compilation unit). The processing of aspect associations happens only
when running a recursion in Process_Decisions_Defer on children of nodes
passed to Traverse_One.

Change-Id: I368e39eed6aff9b11fc8e9be92a4984bf2d03f2d
TN: U520-030