Prevent witness calls from being optimized out by the compiler
With some compilers (only 7.2.2 is known to do so at the moment),
and with optimization levels -O1 and higher, the calls to the witness
functions inserted during instrumentation of declarations is optimized
out because this function is declared in a pure package, and the result
of the function call is stored in a variable that is never referenced
after that.

To prevent this kind of optimization, make the return type of the
witness function volatile.

Change-Id: I9a746293832af5dd89ac7709df9eeaaee064c8f2
TN: U616-021