Remove wrapper package for augmented expression function
This nested package triggers too many freezing bugs in GNAT
and we do not have any compelling evidence that introducing
new primitives is an issue (preliminary testing on GPR2,
SPARK 2014, ALS and QGen do not show any build issues or
semantic issues).

Since the augmented expression function can now be a primitive,
we sometimes need to emit a declaration for it, next to the
declaration of the original expression function, when there is
one. This is to avoid freezing issues by declaring a primitive
after the freezing point.

We do keep the wrapper package for the specific case where the
original expression function is a primitive of its return type,
and that type is tagged. In that case, without the wrapper
package the augmented expression would be a primitive, and we
would need to generate an overriding primitive for every type
derivation (see RM 3-9-3, 6/2)

Change-Id: I5366e25313cc374a27cd897e5871f3f256a19c21
TN: U625-016