SC_Obligations.BDD: add a context handler in Enumerate_Paths
This ticket showed that we can have a Constraint_Error exception raised
in this procedure for sufficiently complex decisions. We have plans to
remove this error for binary traces, but things are not so easy for
source traces, so add a context handler to give users helpful
information about what part of their code triggers the error.

Change-Id: Id34e400f1573bc44ed4efd393a2c1d8e6eee3616
TN: U416-006