Do not count BDD evaluation paths when processing binary traces
For complex enough decisions, there are just too many evaluations paths,
so we get a Constraint_Error exception during the enumeration. This path
count is unused when dealing with binary traces, so do not compute it in
that case.

The problem is still there when dealing with source traces, but there is
no obvious solution to this problem: it is still to be addressed in the
future.

Change-Id: I197cd182ed9f247222faf95e94dd20805bed292e
Depends-On: If323bc10da1c33048cfe9be1690cba66ef0e91ec
TN: U416-006