Fix gnatcov builds
The builds failed because there is a call to Has_Decision in
instrument_ada_unit.adb:4422 that only has a single parameter and
not two as expected by the spec.
There are two concurrent changes that lead to this situation:
I88d567e708a3f6266c2ee3f2611d6c0d518739fc modified Has_Decision and
added an additional parameter, and
I20213011255050f910fc9564869e3e72adf9449d added a new call to
Has_Decision (which only had a single param then).
They were both put on review at the same time, the second commit
was approved before the first one, and individually they each build
correctly.
However, the first commit was not rebased on master once the second
commit was merged, and so the CI did not pick up that there was an
error once the two changes were applied.

Change-Id: I112077a6a81220963ffb78c1dc6e72f5a6f902f4
TN: U721-006