Fix source traces parsing
As the Unit_Coverage_Buffers record type in gnatcov_rts-buffers.ads was
extended with a Project_Name field, to account for the C source traces
model (as a C source file is identified both by its filename, and the
project to which it belongs), the parsing needs to be updated as well.

There were no issues with Ada source traces, as they make no use of this
new field, which is thus always left empty.

Change-Id: I6c680fba4ae6a5085dcb96a7248999c9c3686c69
TN: U208-020