Add option to suppress the emission of a coverage report
This is useful in scripting scenarios, for instance when the report
format is specified in the root project file and the script should
create a checkpoint, without outputing a report (which would be the
case at the moment). This will also be usefull for the
gnattest/gnatcov integration makefile.

Change-Id: If4fc8a6c131b291db936ada27c0cfa57153c2e97
TN: U811-028