Rework the description of decisions/conditions for mcdc
State upfront that only short-circuit operators drive the split
of decisions into separate conditions, which wasn't clear prior to
the change and caught several users by surprise.

Refactor "Decision composition rules for MCDC": Expand examples as
a separate subsection and extract the handling of non short-circuit
operators on its own. Pick easy to grasp titles for each.

Contextualize the examples showing gnatcov run, to mention that this
is specifically for binary traces.


Part of U608-027 (clarify gnatcov doc on short-circuit for mcdc)

Change-Id: I2df52a59402e863740d17886cce352d0013ba8e1