Fix ability to do decision analysis from mcdc instrumentation
Before this change, the info stored in sid/checkpoint files when
instrumenting for stmt+mcdc advertised that only stmt and all the
variants of mcdc are possible at analysis time.

Attempting stmt+decision analysis complained that decision
analysis is not possible, which is incorrect.

This change fixes this by adding "decision" to the set of
criteria for which information is available in a sid/checkpoint
when instrumentation is performed for MCDC purposes.

Part of U527-020 (spurious gnatcov error on level incompatibility)

Change-Id: I063b180920f7c3fb027c79ce38fa812cc12563ab