Replace hardcoded lists of level switches by computation
In the fatal error message emitted on the use of an
object coverage criterion together with --annotate=report,
or of a source criterion together with --annotate=asm.

Reformat the messages a bit to better fit an 80 columns
terminal.

In the course of U623-030 (confusing --help on levels for
gnatcov instrument)

Change-Id: Icdbfa2c5bda06d58bd8fae3769eb8680086e130c