decision analysis: fix false positives errors
When seeing code for a condition attached to a decision which is not the
"current" one on our evaluation stack, the tool assumes that a new
decision evaluation starts. With quantified expressions, now holding a
sub-decision for their predicates, we might legitimately be seeing
remnant code pertaining to a parent decision, upper in the evaluation
stack.

We will resolve this by assuming that evaluating object code that is
supposedly part of the root decision after a sub-decision is not a new
occurrence of the root decision, but is part of it.

Change-Id: I2f84564ccaca283723a37ee5b8b19e9e95794bbb
TN: T310-022