Fix trace dump in testsuite
For cross testing, we use the --dump-trigger=main-end option, as atexit
is not available. We also add a last chance handler, to catch abnormal
terminations of the main program. This does not work for C, so it was
modified to account only for Ada mains, but there was a bug left, which
this commit fixes.

Change-Id: Ib6d5802a96c5be0f387333bb411f3d0257b05249
TN: U208-020