Fix C instrumentation for ravenscar profiles
We use by default the --dump-trigger=ravenscar-task-termination when
instrumenting a program compiled with a ravenscar runtime. But this
dump-trigger do not work for C main.

Arrange for the testsuite to pass the right dump-trigger in such cases
and fallback to main-end with a warning message in gnatcov if such
situation arrises (we won't quit with an error, as this would prevent
instrumentation of a multi-language-drivers project with ravenscar, if
such thing exists).

Also fix regression introduced in
Ic4192e0fb69f4055e2e55708aabdc60a0ac02a23, where adainit was invoked only
for --dump-trigger=atexit, and not for other dump triggers (such as
main-end).

Change-Id: Icf5eb88df588f2a54ef0e39dee20d68e05bc7575
TN: U208-020