C instrumentation: fix initialization of the Ada runtime
As we use the Ada runtime to dump C traces, we need to call adainit
before using it. We were previously calling it in the atexit handler,
which resulted in a segfault for older toolchains. Calling it at the
beginning of the main works and is a better option, so we will opt for
that alternative.

Change-Id: Ic4192e0fb69f4055e2e55708aabdc60a0ac02a23
TN: U208-020