Minor rework of gnatcov doc figures processing
Introduce a more systematic automation of the dot to pdf/png
conversions for lone figures, to facilitate the addition of new
figures.

Rename existing lone figure files so they feature a fig_
prefix, allowing their handling with GNU make patterns not
encompassing the family of files used for to illustrate
the effect of -P/--projects on project files selection.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: I4d302b5ebda023f23c7ebc62230111d22d97e1a0