Rework the gnatcov "Concepts" intro section
Expand the outline of the framework very high level traits
as a list instead of a long sentence.

Reorder paragraphs to present source and object
covrerage in sequence, then the main notion of SCOs
and units of interest just before a figure illustrating
the processes.

Reword (clarify) the notion that we don't support mixing
trace kinds, leveraging the just introduced process diagrams.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: I16cc2233e2d5b6cd6c8fdeba2c79381c6d7bdffe