Refactoring to simplify gnatcov driver dependencies
This commit moves around some code, to avoid in the end having a
dependency to the instrumentation source files in gnatcov.adb. As
instrumentation brings a dependency to both libadalang and libclang, it
makes the executable quite huge.

The notable changes are in inputs.adb (avoiding the dependency to the
unit diagnostics), and in the coverage / coverage_options units. We move
the declaration of coverage level types and utilities functions from the
unit coverage to the new (dedicated) unit coverage_options. This cascades
to adding a dependency to coverage_options in various units.

The size of gnatcov driver in dev mode goes down to 16M (it was 220M
before).

Change-Id: Iff87cc48764f25ea025d251ee6480ebef614849c
TN: U917-007