Remove dependency to langkit / libadalang in gnatcov32
This refactors the instrumentation sources of gnatcov, to avoid bringing
an unneeded dependency to langkit / libclang when using the stubbed
version of the Instrument unit (as gnatcov32, which does not provide
instrumentation, uses this stubbed version). The gnatcov32 executable
size in dev mode goes from 135M to 87M.

We also move some declarations where they should be, notably part of the
code in the Instrument.Common unit that was specific to Ada
instrumentation.

Change-Id: I9361a37809cee5c9420caaa1f704c9d12548f293
TN: U917-007