gnatcov decision analysis: remove superfluous warnings
This commit arranges to remove some gnatcov warnings that were false
positives in the decision analysis sense. The decisions were rightly
analyzed, but unconsistencies in the debug information (notably for
finalization code) resulted in gnatcov warnings (gnatcov thinking it
was reanalyzing a decision after it had left it).

Change-Id: I54afc7bda13b4bc7f2deb0e38ed434d57b4a21fe
TN: T917-015