Simplify the "supported configurations" gnatcov doc sections
Merge and simplify (a lot) the two sections on "supported configurations"
and "selecting a trace mode", the points of which were interleaved and a
bit confused.

Present the support information as a table instead of as a
sequence of text paragraphs.

Take the opportunity to update the information about what
we support and not, which changed significantly since 21.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: I960915715c9c8f58934e7d95d1663dcbcc3607d6