Refactor the "process overview" gnatcov intro sections
Merge what were the "process overview" and "example sequences" sections
and present the result explicitly as an introductory
illustration without details, only intended to provide a feeling
of what the operations actually look like.

Remove the attempt at summarizing the general interface of all
the commands, redundant with what other chapters do, too easily
obsolete and not really useful at this spot.

A list of links to the relevant chapters will be provided as
part of a rework of the "going further" section, in a forthcoming
commit.

Introduce the typographical conventions that were described there
in a distinct section instead, at the end since we don't use them
any more in this intro.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: Ie682a4cbcb9054536172a39129f75f743156b828