Rework the "going further" gnatcov doc intro section
Introduce the chapters of interest in a more straightforward manner,
providing a list of links to each of them.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: I5e8cd891e60c66040f8a8f30d8f59bfc75f9fd06
Rework the "going further" gnatcov doc intro section

Introduce the chapters of interest in a more straightforward manner,
providing a list of links to each of them.

Part of U211-004 (clarify gnatcov introduction doc)

Change-Id: Ib5d38c814108a01df732717a2bf0f6c588b129e6