Base the decl for augmented EF on the one of the original EF decl
Before this patch, when inserting a previous declaration for an
augmented expression function, the subprogram specification in the
declaration was a copy of the subprogram specification in the
actaul expression function. This does not work when the visibility
of some type is modified by use-clauses in between the original EF
declaration and its completion.
In the following example,

   with Pkg;
   package Bar is
      function Foo (Self : T) return Pkg.Other_Type;

   private

      use Pkg;
      function Foo (Self : T) return Other_Type is (...);
   end Bar;

Foo was be instrumented as
   with Pkg;
   package Bar is
      function Foo_With_State
        (Self : T, Dummy : MCDC_State_Holder) return Other_Type;
      function Foo (Self : T) return Pkg.Other_Type;

   private
      use Pkg;
      function Foo_With_State
        (Self : T, DUmmy : MCDC_State_Holder) return Other_Type is
        (...);
      Function Foo (Self : T) return Other_Type is
        (Foo_With_State (...));
   end Bar;

This change fixes it by basing the declaration for the augmented
expression function on the declaration of the original expression
function.

Change-Id: Ie856c91f2cc7cd81bdb2f8ee29beed4f38516e1b
TN: UA11-007