Introduce a C version of the runtime
The Ada core runtime is replaced with an equivalent C version, and the
Ada version simply consists of thick bindings to the core C function.

This results in simplifying a bit the C instrumentation process (as we
previously had to bind to the Ada runtime), and enables us to generate
only C files to instrument C code (and we make sure we keep the symetric
property, i.e. to generate only Ada files when instrumenting Ada code).

This does not change the installation process, i.e. no user breaking
changes are introduced by this commit.

Change-Id: I2e1eefef6152a8669cb9e5b1830b9147a70aca7b
TN: U208-020