decision map: do not push incomplete decisions on the stack
As debug information is not 100% reliable for our decision analysis,
gnatcov sometimes think it is analyzing decisions when it is
encountering a branch for which the debug information points to the same
source location of a condition.

We do not want to trigger decision analysis when encountering such
branch, as they will result in incomplete decision analysis and tons of
spurious warnings. To know if some parts of the decision analysis is
wrong or missing, we shall rely only on `CONDITION lacks edge` warnings.

Change-Id: I1136d10ef1ae8324309b0235abef60ba5152b501
TN: UA06-015