Fix binary traces for aarch64-elf using runtime embedded-zynqmp
This is related to the transition to gcc 11, and notably changes in the
debug info format. In DWARF file tables, the current compilation file
(and the current compilation directory), are represented with index 0.

This causes source file conflicts in gnatcov files table, notably when
relative paths come to play, which is the case for embedded as we link
against newlib, and there are relative paths for a lot of source files
coming from here, see below.

 The Directory Table (offset 0x2ef8, lines 5, columns 1):
  Entry	Name
  0	(indirect line string, offset: 0x7df): \
     /it/sbx/wave/aarch64-elf-linux64/newlib_bootstrap/\
      build/aarch64-elf/newlib/libc/stdlib
  1	(indirect line string, offset: 0x865): \
     ../../../../../src/newlib/libc/stdlib

 The File Name Table (offset 0x2f12, lines 7, columns 2):
  Entry	Dir	Name
  0	0	(indirect line string, offset: 0x99d):\
     ../../../../../src/newlib/libc/stdlib/mallocr.c
  1	1	(indirect line string, offset: 0x9c3): mallocr.c

gnatcov thinks the source files are different, and inserts them both in
the files table. At some point, when it freezes the files table, it can't
generate a unique name for both files and errors out.

We workaround it by simply arranging to always pass full paths for
directories name (as noticed above, the first directory path is full while
the 2nd is relative). This makes filenames consistent (we will get the same
ones when combining the needed directory with each filename).

Change-Id: I5f804ffd8d137c69a4209bfd5a7ab4539e6553b0
TN: UA25-056