Robustify dwarf5 file/directory names processing
A few changes which, combined together, let the tool process
gcc-11 binaries on Windows more dependably:

- In Is_Absolute_Path, recognize '/' as a possible dir separator
after the drive name in Windows paths, in addition to '\'.

- Simplify the processing of dwarf5 file_names attributes.
Make sure we never compose a dir indication with an already
full path filename.

- Assimilate Canonicalize_Filename to OS_Lib.Normalize_Filename,
which does some of the things we attempted better and a bit more.
Let it deal with the necessary differences on Windows vs Unix,
which will stop lower casing all characters in Windows file names.

- Use Canonicalize_Filename consistently in traces_elf.

Part of UB03-055

Change-Id: Ibfa6f840b1dfc75fba717fcaec3dadd313d66269