Ignore inline subprogram address ranges when doing object coverage
In order to store address ranges, we use a complex tree structure that
makes the assumption there are no address ranges overlapping and  non
nested. This does not seem to be the case for rust generated object
code. For the purpose of doing object coverage, we do not need to have
that information, so we will simply ignore it for now.

Change-Id: I32e45df94cfe9041754c5cb9f63d9aaf1196f5a6
TN: UA28-035