Move gnatcov doc section on source-search/rebase
Move the gnatcov doc section describing source-rebase
options after the set of sections describing the various
formats, so we have a first block of sections just describing
the formats, then sections describing notions of interest for
subsets of the formats (e.g. those producing annotated
sources).

This prepares for the addition of a new section of the
latter kind.

Part of UA20-021 (document synthethic metrics in gnatcov reports)

Change-Id: If9807bb4302488c58fb3bc35d507ede6e2c8102e