Document and add warning for unsupported instrumentation cases
The instrumented code for expression functions which are primitives
of their tagged return type was so far potentially missing overriding
primitives. This commit disables their instrumentation.
Recursive expression functions which are primitives of some tagged
type may also introduce freezing issues if the EF is not the last
primitive in the declarative region, so gnatcov will no longer
instrument them either.

Change-Id: Ide21b07e8f3001c5ccfbd2ae479e5cf39b3c2cdb
TN: U625-016