Add possibility to produce multiple reports in one execution
Add the possibility produce more than one report per invocation
of gnatcov. This is done by passing a comma separated list of
annotation formats to --annotate, or by including the --annotate
option multiple times on the command line. This is a backward
incompatible change to the behavior of --annotate.

Change-Id: I47f477e7c8f041cf3466849f84a03d2f613bb238
TN: V201-010