Rename gnatcoverage documentation and paths
In order to avoid confusion between the product and the package, the
gnatcoverage package is renamed. In order to follow these changes, we
renamed the documentations and the install paths for them, both in anod
and in the gnatcoverage repository. Also, remove of the temporary
duplicata between gnatcoverage and gnatdas.

TN: V127-033
Change-Id: Ic13bbb5587848fbf704efb5f5c29b4d1b888ac5d
Depends-On: I50cd724130f68d372c36db410e15d8cc20d0a137