Fix soundness issue in the decision map analysis
gnatcov was incorrectly labeling some control flow branches as outcome,
when they were branching into an inlined body.

In the following example, assuming Orelse is inlined:

function Orelse (B, C : Boolean) return Boolean is
begin
   return (if B or else C then True else False);
end;

function F (A, B, C : Boolean) return Boolean is
begin
   return (if A and then Orelse (B, C) then True else False); -- #D1
end;

gnatcov incorrectly labeled an edge as implementing the valuation `False`
for the condition A, and as an outcome of the decision #D1, which is
unsound behaviour. As it could not find an edge for the valuation `True`,
it also resulted in a `condition lacks edge` warning.

TN: T706-019