Enhance qualif testsuite for complex if expression
As noticed in the qgen codebase, when we have an if expression inside a
decision and if the operands of that if expression are complex boolean
operands, gnatcov failed to find edges for all of the conditions.

This is because the if expression itself (as an operand of the top-level
decision), and the operands of the if expression are conditions, but they
share the same control flow branches.

For example, if we take a look at this example:

(if A then B and then C else D and then E) or else B

If we focus on the valuation of the C condition

Condition  Val     Res   |    Condition  Val    Res
---------------------------------------------------
(if_expr)                |    (if_expr)
C          ------> True  |    C         -------> B
            True         |               False

Here, we see that the if expression "condition", and the C "condition"
share the same branch.

To work around this, the compiler now introduces a temporary holding the
value of the if expression.

Change-Id: If37d8261c99a1ce57c61dbe62e5decb28d2b9bb6
TN: U709-021