gnatcov_rts_c-traces-output.c: fix signedness issues
With extra warnings, compiling this unit yields warnings such as:

    pointer targets in passing argument 2 of 'write_bytes' differ in
    signedness

This is because we try to convert a char* to a uint8_t*, whereas on some
platforms char is signed (uint8_t is always unsigned). Use void* where
appropriate when we really mean "pointer to anything".

Also fix signedness issues in FOR loop exit conditions.

Change-Id: Ib7682df957562a15f637985337af158710fb2b02
TN: U208-020