GNATcov_RTS: enable all C and Ada warnings
This commit enables all warnings when compiling GNATcov_RTS to detect
potential issues in the code base. Since this project is supposed to be
built with a very wide range of compilers, we cannot know for sure that
all warnings are addressed, so do *not* process warnings as errors
(-Werror for C/-gnatwe for Ada).

Change-Id: Ibb936b82ab474b31db05eb372b965d8cb0726ad4
TN: U208-020