Produce a valid unit name for the buffers list unit
If the root project is a child project (e.g. p-child.gpr), we can't use
the name of the project file "as is" to produce a unit name for the
buffers list unit (e.g. gnatcov_rts-buffers-lists-p-child.ads). We don't
want to create the whole package hierarchy there (e.g.
gnatcov_rts-buffers-lists-p.ads), so to keep things simple, we'll use a
unique slug of the project name.

This always was a problem, but as the generated buffer list unit was not
with-ed in any other unit before 21cead26ff19d1b1f5cebb792dcf916a58afcec9,
we did not get compilation failures.

Change-Id: Id8b12612f7a2ff990a4691397625ec30322d6514
TN: U208-020