Check if the file descriptor is valid instead of checking errno value
Failing calls can set errno to some non-zero valees, but successful calls
don't set errno to 0. To check an I/O error, we will check the validity
of the file descriptor (i.e. not null), and if it is invalid, rely on
errno to get a proper error message.

Change-Id: Ie78254f1cb78da28a1beb69191b2160a79958426
TN: U208-020