Turn backslashes into forward slashes for --source-rebase
This makes it possible to use --source-rebase to locate Windows absolute
file names on a Unix system.

Testing this requires adding pre-produced binary files (executables or
checkpoints produced both on Windows and Unix systems) into the
testsuite, which adds maintainance burden. The testcase for U204-026
will need to do that anyway, so this new testcase will test this new
option.

Change-Id: I27b38e6381234f79ed9a80c36c563f720e7bcbbc
TN: V318-007