Mention SID files in --subdirs' help message
This option makes the project handling machinery to automatically change
the object directory of all loaded projects, so it also affects
"gnatcov instrument", making it load SID files in that subdirectory.

Note that it also changes the directory in which instrumented sources
are created, but unlike SID files (which are part of gnatcov's user
interface), this is an implementation detail.

Change-Id: I85c3e0de8ea3210c7ce397ccac27b87466caafef