System_Commands: do not modify gnatcov's own environment variables
This commit replaces uses of GNAT.Expect in System_Commands with the
GNATCOLL.OS.Process API. The latter allows gnatcov not to modify its own
environment in order to pass environment variables to the subprocess.
Also tweak a bit System_Commands' verbose output.

This fixes no known bug for now, but is nonetheless a code smell, which
could trigger bugs in the future, for instance when running multiple
commands from the same gnatcov process.

Change-Id: I39155a0260ba4d3caed2ce167fc40622130a2072
TN: U211-014