Document 7.x.2 limitation with expression functions and optim
Not all optimization issues could be resolved with 7.1.2
(up to 7.3.2), in particular regarding coverage of expression
functions.

Given that support of Ada 2012 constructs was still limited for
these releases, and that the customer reporting the original
optimization issue did not use any expression functions, simply
document this as a tool limitation at this stage.

Change-Id: Idbf336504eb2779ec20dab72f2042705b7bed140
TN: U616-021