Setup: rework project handling
So far, "gnatcov setup":

* optionally took the directory that contained the instrumentation
  runtime project to build and install (--runtime-source-dir), plus the
  name of the project to install (--runtime-project, to allow concurrent
  installs of the same project);

* copied that directory in a temp. dir. and ran a regular build on the
  copied project file.

This is brittle for several reasons:

* it relies on string substitution to rename the project file, which is
  a hack that does not work in corner cases which will be surprising for
  users (user modified project files in which the casing varies, or that
  contains string literals, which must not be changed);

* the UI is not very user friendly: need to pass the containing
  directory rather than the actual project file to build, GPR problems
  with homonym projects if users want to extend the standard runtime
  rather than just copy+modify the one shipped with gnatcov, ...

This commit changes the UI so solve these problems. After this commit,
"gnatcov setup":

* optionally takes the instrumentation runtime project directly
  (positional argument), plus the desired installation name
  (--install-name);

* does not post-process the given project file; instead, it creates an
  extending project when the requested installation name does not match
  the given project, runs gprbuild on that project instead.

Both the concepts and the UI should be simpler to document and use in
practice, and this way to build the instrumentation runtime should be
more robust.

Note that one problem remains with this new scheme: the copy of the
directory that contains the instrumentation runtime can be costly (for
instance if users left big binaries there unrelated to the work of
"gnatcov setup"), and that directory may not be self-contained (for
instance if one of its source directories lives in a parent directory).

Ideally, an out-of-tree build would remove the need to do the copy and
would avoid breaking the layout of projects that spread to parent
directories, but there currently isn't a GPRbuild feature to do this
exactly.

Change-Id: I86383f4b45762f08470ab6625e1da6d812a9c306
TN: U211-014