Instrument: save the dump channel in the root project object dir
Make "gnatcov instrument" write in the object directory of
the root project a JSON file that describes the selected dump channel
for the instrumented mains.

This allows tools that must work on the source traces to automatically
detect where to find them: either in a *.srctrace file, or as base64
encoded traces in the main's standard output. This will be very useful
in IDE integration.

Note that this information is not the definitive way to determine where
to look for source traces in general: even knowing all --dump-* options
used during the main's instrumentation is not enough to know where to
look for source traces: for instance in the case of
--dump-channel=bin-file, the source trace file can be produced in a
location determined at run-time (with an environment variable, for
instance), so this information is just a hint for these tools.

This patch also modifies the testsuite to use this information: check
that the dump channel described in the JSON file is the one requested,
or to guess the dump channel when the testsuite does not have this
information.

Change-Id: I4f973ec509f67bbf7140ced2b4f2b648a37da5ee
TN: V309-017