Rework the C instrumentation process to keep macro expansion information
To have interesting information about SCOs in macro expansions, we need
to parse the original source file before instrumenting it.

As instrumentation must operate on the preprocessed file, introduce a
pass that runs before instrumentation, that is dedicated to emitting the
SCOs (to have the right column numbering) and filling in macro expansion
information.

Change-Id: I43fd7a9b729aacd00f9a01d6eb34b177e6f7ae48
TN: V426-009