Setup: fix automatic uninstallation
Do not pass --prefix when there is no prefix. We used to pass
"--prefix= gnatcov_rts" in that case, and gprinstall interprets this as
"the prefix is gnatcov_rts" (whereas it is supposed to be the name of
the project to uninstall).

Also convert the project name to lower case, as it is case-sensitive for
gprinstall, and must correspond to the case of the project that is
installed.

Change-Id: If320f6626b43165ed9741ab6ed5760d0d55d463e
TN: U211-014