Instrument.C: always fill the decision location when computing SCOs
We used to fill the control location, for a decision inside an IF or a
WHILE. The precise sloc of the decision was filled afterwards, when
computing high level SCOs.

As this control location is only needed in the context of binary traces,
and as we need the precise location when we are computing low level SCOs
for C instrumentation, always pass the location of the decision (even for
decision expressions).

Change-Id: I676cfed62d5f1d80f99f8957fa35a68bd3dc34cd
TN: V426-009