Tweak the instrumentation runtime to make it compatible with AAMP
The AAMP target comes with a notable feature: a storage unit size of 16
bits. As we rely on the binary layout of intermediate structures to dump
the traces in our instrumentation runtime, we specify representation
clauses that are wrong (and uncompilable) for this target.

To deal with this target, we tweak the code of the runtime to make it
resilient to a 16-bit storage unit size. We do that by decorrelating
the trace writing from the in-memory representation of records / strings.

We also remove the use of discriminant in records, as their implementation
(and computed allocation size) appears to be quite fragile in the compiler
implementation. We switch the use of discriminant-length defined strings,
to System.Address in the record types of the runtime (notably for the
Unit_Coverage_Buffers type).

Change-Id: I75499a839b6ed718af14693f64a49acd157fff04
TN: V404-035