Add a custom runtime project file for the AAMP target
As AAMP does not support libraries, make a non-library project file for
this target. Also remove the C source files from this custom runtime, as
there is no C compiler for this target (and the C source files only
provide a C version of the witness subprograms, which won't be needed for
AAMP).

Change-Id: I0feef3f38480016733c6c8f22d08ef24bd406cb8
TN: V404-035