Instrument: fix the handling of decls with multiple defining identifiers
A recent update in Libadalang (intentionally) created new exceptions
when querying the "ghost" status for declarations that define multiple
entities, such as "A, B : Integer", as one entity may be ghost but not
the other.

Query the ghost status directly on each defining identifier node instead
of on the declaration node to avoid this problem. This fixes the
testsuite in --all-warnings mode.

Change-Id: I727c15e8f814f3e8586e89b2040e227710c8f88c
TN: V517-019