Compute line coverage state only once to avoid duplicate messages
Ticket V201-010 implemented the possibility to produce multiple
reports in a single "gnatcov coverage" execution. With this change
the procedure Compute_Line_State from Coverage.Source is called
once for each line, for each annotation format, which results in
the violation messages being emitted multiple times.

This change adds a flag in the Line_Info record to indicate that
the line coverage state has already been computed so that
subsequent calls to Compute_Line_State do not recompute the
coverage state.

Change-Id: Ic872c284420def78b5fb8336a63969f3ec65bcd5
TN: V520-019