Convert qualification scripts to e3/python3
Add parens on calls to "print" in a few python scripts.

In the STR/Makefile, temporarily for the invocation of
sphinx-build, prepend the path to sphinx-build on PATH
to have the associated python distrib ahead as it exposes
specific packages that sphinx-build uses.

In STR/genrest.py, adjust the expected texts for test
statuses to those emitted by e3. Remove trailing whitespaces.

Part of V523-030 (gnatcov qualkit with e3/python3 on lima)

Change-Id: I58c039d3058f88c5de75bff0add0a0c9a9deb115
Depends-On: Id7de2b90cd84c997cbbcb382798268caa1e85014