Fix behavior of the --post-run-cleanup option
The --post-run-clean option should clean up iff the status of the test is
FAIL. As the status was still uninitialized (to ERROR) when we were
checking whether we should clean up the test directory or not, it
resulted in always cleaning it.

To fix this, we analyze the result of the test (where the test status is
filled), before tearing down (where the cleanup occurs).

TN: V523-030
Change-Id: I333537009a8db12b28e0c66abccdd869ba6c0578