Tigthen doinstall
Within doinstall, cleanup possible remnants of a gnatcov_rts
previous installation before untar'ing, as we want the new installed
runtime dir to contain only the new sources.

When generating the doinstall script, fix the path to the
"gnatcov" program used to figure out the tool version, to
designate the one we're actually installing. We were using
"gnatcov", which could result in "no such file or directory"
or resolve to a random version depending on the current
PATH.

Part of V524-022 (gnatcov installer cleanup coverage runtime dir)

Change-Id: I011e78626977d6ab3ea5926f418e78f1036e42ce